requirements.txt
# Top-level requirements for local development.
# Install into a venv: python3 -m venv .venv && source .venv/bin/activate && pip install -r requirements.txt

# Detector dependencies
dnspython>=2.4.0
httpx>=0.25.0

# Shared detector/attacker dependency
scapy>=2.5.0

# Testing
pytest>=7.0.0
